requirements.txt
Cython==0.24.1
Django==1.10.1
Fabric==1.10.2
Pillow==3.2.0
PyDispatcher==2.0.5
Pygments==2.1.3
Flask==0.11.1
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.1.9
Scrapy==1.2.0
Twisted==16.4.1
Unidecode==0.04.19
Werkzeug==0.11.11
appdirs==1.4.2
appnope==0.1.0
arrow==0.8.0
attrs==16.2.0
beautifulsoup4==4.5.1
cffi==1.8.3
click==6.7
contextlib2==0.5.4
croniter==0.3.14
cryptography==1.5.2
cssselect==0.9.2
decorator==4.0.10
dicttoxml==1.7.4
django-annoying==0.10.3
django-bootstrap-breadcrumbs==0.8.2
django-carton==1.2.1
django-ckeditor==5.3.0
django-compressor==2.1.1
django-cors-headers==2.0.2
django-debug-toolbar==1.6
django-extensions==1.7.4
django-extra-views==0.8.0
django-filebrowser==3.9.1
django-filter==0.15.3
django-haystack==2.5.0
django-hitcount==1.2.2
django-htmlmin==0.10.0
django-mptt==0.8.6
django-mptt-admin==0.4.1
django-password-reset==0.9
django-postalcodes==0.4.0
django-registration-redux==1.4
django-rq==0.9.4
django-sass-processor==0.5.2
django-tinymce==2.4.0
django-treebeard==4.0.1
django-widget-tweaks==1.4.1
django-yandex-kassa==0.9.12
djangorestframework==3.5.3
djangorestframework-filters==0.9.0
ecdsa==0.13
et-xmlfile==1.0.1
future==0.16.0
geoip2==2.4.2
geos==0.1.0
gunicorn==19.6.0
html5lib==0.999999999
ipython==5.1.0
ipython-genutils==0.1.0
itsdangerous==0.24
libsass==0.11.2
idna==2.1
jdcal==1.3
lorem-ipsum-generator==0.3
lxml==3.6.4
maxminddb==1.2.1
openpyxl==2.4.5
paramiko==1.16.0
parsel==1.0.3
pbr==1.10.0
phonenumbers==7.7.1
psycopg2==2.6.1
pyOpenSSL==16.2.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.16
pycrypto==2.6.1
pycurl==7.43.0
pysolr==3.5.0
pysxgeo==0.0.3
python-dateutil==2.6.0
python-slugify==1.2.1
pytz==2016.7
queuelib==1.4.2
redis==2.10.5
requests==2.11.1
rq==0.7.1
rq-scheduler==0.7.0
service-identity==16.0.0
six==1.10.0
sorl-thumbnail==12.3
sqlparse==0.2.2
stevedore==1.17.1
transliterate==1.8.1
virtualenv==15.0.3
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
xlrd==1.0.0
xlwt==1.3.0
xmltodict==0.10.2
yandex-money-sdk==0.1.3
zope.interface==4.3.2
webencodings==0.5
w3lib==1.17.0
wcwidth==0.1.7
traitlets==4.3.1
solr==0.4
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
proj==0.1.0
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pykml==0.1.0
pyparsing==2.2.0
raven==6.0.0
simplegeneric==0.8.1
git+git://github.com/gadjimuradov/imbox.git
